Import the GUI code + hook it up to the revised stitcher (#6)
This commit pulls in the code from

https://github.com/sohamazing/image-stitcher/blob/622b06e5a9151c35453ea03ffc9e1b607547159c/stitcher_gui.py
with some minimal modifications to hook it up to the new stitcher
object.

The main difference here from the original is that I've added all the
GUI dependencies as optional (in a group called `gui`) so that the CLI
stitching can be used without having to worry about having QT and all
its dependencies installed correctly, but it's still easy with one extra
flag to pull in the GUI dependencies. Otherwise, I've mostly left the
code as-is. I'm not a QT expert, so I don't have a good idea of what if
any cleanup there is to do.

Tested by:
- `./dev/format.sh`
- `./dev/autofix_lint.sh`
- `./dev/type_check.sh`
- `uv run --extra gui python -m image_stitcher.stitcher_gui`
- see that the GUI starts up
- for both formats:
  - load up the folder of images I used for the CLI testing
  - do the stitching
  - open the result in napari, and verify it looks as expected