Choose the chunk size for the output dask array based on available memory (#10)
While experimenting with using zarr for arrays instead of dask, I
discovered that we could get a modest extra speedup for dask by using
even larger chunks. This commit makes us use chunks up to the amount of
available memory (up to a maximum of 2**31-1 bytes, which was a hard
limit for zarr, and when testing with larger chunks on dask it seemed to
hang).

Tested by:
- `./dev/format.sh`
- `./dev/autofix_lint.sh`
- `./dev/type_check.sh`
- `./dev/run_tests.sh`
- run a manual test on an 8x8fov, 3000x3000px, 5-channel image, see it
completes faster than before.