Remove a bunch of unused code for non-implemented features (#16)
Summary:

The original code on which this stitcher was based had a number of extra
features like flatfield correction, registration, and assembly of
higher-level zarr files, which are not implemented here. I've removed
the code and the parameters from the GUI that aren't hooked up to avoid
confusion. (If these get added back later, the code can always be
brought back from git history.)

Tested by:
- `./dev/run_tests.sh`
- `./dev/autofix_lint.sh`
- `./dev/type_check.sh`