Add setup and install scripts for the GUI (#17)
For ease of use, we'd prefer not to have people run a bunch of terminal
commands manually. In contrast to the main microscope software, we'd
also like to keep this tool isolated in a virtualenv to reduce the
chance that analysis tools can break the microscope software.

This commit adds some setup scripts that can help do this:
- `setup_ubuntu_22_04.sh` there's a bunch of extra system dependencies
we need on ubuntu (mostly for QT); this installs them automatically. I
chose ubuntu 22 since it's the same OS the microscope software
installation script targets.
- `dev/ensure_uv.sh` a wrapper around the installer for uv, which we use
to manage the virtualenvs, python installation, and python dependencies
- `run_gui` ensures that uv is installed by calling into
`dev/ensure_uv.sh` and then runs the GUI command. This is also
double-clickable on mac (and presumably linux, though I don't have a
non-containerized ubuntu 22.04 installation to test on)

Tested by:
- create a fresh ubuntu 22.04 container
- `./setup_ubuntu_22_04.sh`
- `./run_gui`
- see these succeed, I get the stitching gui, and can stitch some images
- double click `run_gui` on mac, see I get the stitching gui and can
stitch some images