feat: add flatfield correction (#21)
@cjfuller implemented the actual functionality here - doing flatfield
correction. It is added both as a cli option, and a checkbox in the gui.

The use of basicpy caused dependency issues, though. I was able to get a
setup with just `pip` working by removing most of the constraints on the
packages (I think the numpy constraint might have been the major
issue?), but replicating this with uv still resulted in build issues.

I also downgraded the python requirement to 3.10 since that's the
default on ubuntu 22.04 lts.

Since customers seem more familiar with conda, Hongquan asked to go that
route. So this now installs conda, creates a conda environment, and sets
it up to work for both the cli and gui. The README should reflect this.

Tested By: Running the setup script in an ubuntu:22.04 docker container
and making sure the cli worked. Also, running the same on my own
computer and making sure both the cli and gui work (both with and
without flatfield correction). I also made sure the tests pass, and all
the dev scripts run.

Since we know the issue here were dependencies, it'd be good to have a
few others test.

---------

Co-authored-by: Colin Fuller <colin@cjf.io>