fix: make sure to re-chunk pyramid levels and dask array top level (#22)
For zarr output, we need "regular chunks". Where that is defined as "all
chunks in a dimension are the same size, except for the last that can be
smaller".

We were rechunking in 1 case before, but weren't rechunking for dask
array case and weren't rechunking for any of the pyramid levels. This
adds rechunking to our pre-computed chunk size for all of these.

There is still 1 question mark about the zarr case that isn't handled,
but I didn't have a dataset to test that with.

Tested by: A new unit test that does pyramid creation. It fails before
this PR's content, and passes after.